fix permis

+++ app/Http/Controllers/ListaController.php
@@ -20,7 +20,7 @@ class ListaController extends Controller
      */
     public function index()
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
         $listas = Lista::all();
         return view('listas/index',compact('listas'));
     }
@@ -32,7 +32,7 @@ class ListaController extends Controller
      */
     public function create()
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
         $lista = new Lista;
         return view('listas/create', compact('lista'));
     }
@@ -45,7 +45,7 @@ class ListaController extends Controller
      */
     public function store(Request $request)
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
         // Validações
         $request->validate([
             'description'     => 'required',
@@ -78,6 +78,7 @@ class ListaController extends Controller
      */
     public function show(Lista $lista)
     {
+        $this->authorize('authorized');
         return view("listas/show",[
             'lista' => $lista,
             #'titles' => Mailman::getArchive($lista)
@@ -93,7 +94,7 @@ class ListaController extends Controller
      */
     public function edit(Lista $lista)
     {
-       $this->middleware('can:admin');
+        $this->authorize('admin');
        return view("listas/edit",compact('lista'));
     }
 
@@ -106,7 +107,7 @@ class ListaController extends Controller
      */
     public function update(Request $request, Lista $lista)
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
 
         $request->validate([
             'description'     => 'required',
@@ -139,14 +140,14 @@ class ListaController extends Controller
      */
     public function destroy(Lista $lista)
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
         $lista->delete();
         return redirect('/listas');
     }
 
     public function mailman(Request $request, Lista $lista)
     {
-        $this->middleware('can:admin');
+        $this->authorize('admin');
         if($request->mailman == 'emails') {
             Mailman::emails($lista);
             $request->session()->flash('alert-success',"Lista atualizada com sucesso");
